+++ server/src/main/kotlin/org/kryptonmc/krypton/command/KryptonCommandManager.kt
@@ -4,6 +4,9 @@ import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import kotlinx.coroutines.launch
 import me.bardy.admiral.literalArgument
+import net.kyori.adventure.text.Component
+import net.kyori.adventure.text.format.NamedTextColor
+import net.kyori.adventure.text.format.Style
 import org.kryptonmc.krypton.api.command.Command
 import org.kryptonmc.krypton.api.command.CommandManager
 import org.kryptonmc.krypton.api.command.Sender
@@ -38,6 +41,12 @@ class KryptonCommandManager : CommandManager {
     override fun register(vararg commands: LiteralArgumentBuilder<Sender>) = commands.forEach { register(it) }
 
     override fun dispatch(sender: Sender, command: String) {
-        dispatcher.execute(command, sender)
+        if (dispatcher.execute(command, sender) != 1) sender.sendMessage(DEFAULT_NO_PERMISSION)
+    }
+
+    companion object {
+
+        private val DEFAULT_NO_PERMISSION = Component.text("You do not have permission to execute this command!")
+            .color(NamedTextColor.RED)
     }
 }