+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.1] - 2021-03-12
+### Added
+- Plugin loading! Plugins will now be loaded if their JARs are placed
+  inside the `plugins` folder!
+
 ## [0.11] - 2021-03-12
 ### Added
 - Brand new Krypton API! Still a work in progress, but the core of it is